requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
httpx
xgboost
pandas
scikit-learn
numpy